fix: make variables appear again
It appears that the previous code made enough superfluous workspace and
toolbox refreshes to conceal this problem. Variables are defined on the
workspace, so we need to refresh the toolbox after refreshing the
workspace so that any new variable definitions (or lack thereof) are
picked up and the toolbox code can correctly populate the variables
category.

I'm trying to be neutral, but the state of blocks.jsx is truly saddening
--it seems that every time an issue with blocks refreshing incorrectly
appeared, people just added another workspace/toolbox refresh on
top of it, rather than actually looking into the root cause. Even worse,
after it became apparent that excessive refreshes were causing
performance issues, a mechanism was added to *dynamically check
whether we'd already requested a refresh* rather than take a step back
and ask why all these refreshes were even occurring. By peeling back the
kludges, we can slowly but surely fix the underlying issues.